Move c2d protocol handling into common sdklib
