+++ src/avnet/iotconnect/sdk/lite/__init__.py
@@ -3,4 +3,4 @@ __version__ = '0.1.1'
 from .error import DeviceConfigError
 from .config import DeviceConfig
 from .client import Client, ClientSettings, TelemetryRecord, Callbacks
-from .c2d import *
+from avnet.iotconnect.sdk.sdklib.c2d import *
